requirements.txt

# Core MCP and async dependencies
mcp>=1.0.0
asyncio-mqtt>=0.16.0
asyncio>=3.4.3

# Proxmox API client
proxmoxer>=2.0.0
requests>=2.31.0
urllib3>=2.0.0

# Web framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Configuration and environment
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database and caching
redis>=5.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# Data processing and validation
pydantic>=2.5.0
marshmallow>=3.20.0
jsonschema>=4.20.0

# Utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0
python-dateutil>=2.8.2

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocs-mermaid2-plugin>=1.1.0
